tecnica de integração/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
requests==2.32.3
python-slugify==8.0.4
odfpy==1.4.1
